requirements.txt
streamlit
python-dotenv
requests
pytesseract
pdf2image
PyMuPDF

